New folder/Final-ML/requirements.txt
# Data Science & ML Core
pandas
numpy
scikit-learn
xgboost==1.7.6
lightgbm

# Hyperparameter Tuning
optuna
optuna-integration[xgboost]

## Experiment Tracking
clearml

# Web App & Visualization
streamlit
matplotlib
seaborn

# Utilities
pyyaml     # Dùng để đọc/ghi file .yaml (lưu metrics, params)
openpyxl   # Bắt buộc phải có để pandas đọc file .xlsx
skl2onnx 
onnxmltools